backend/food_service/requirements.txt
flask
pymongo
kafka-python
grpcio
protobuf
